Implemented the detail page

+++ app/src/main/java/com/bangkit/bahanbaku/core/domain/model/RecipeDetail.kt
@@ -7,7 +7,7 @@ import kotlinx.parcelize.Parcelize
 data class RecipeDetail(
     val author: String,
     val portion: Int,
-    val rating: Int,
+    val rating: Double,
     val description: String,
     val title: String,
     val steps: List<Step>,
